requirements.txt
absl-py==0.1.10
aiodns==1.1.1
aiohttp==2.3.9
aniso8601==2.0.0
async-timeout==2.0.0
attrdict==2.0.0
Babel==2.5.3
bitcoin==1.1.42
bleach==1.5.0
blinker==1.4
cchardet==2.1.1
ccxt==1.10.815
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
ConfigArgParse==0.12.0
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
cytoolz==0.9.0.1
dash==0.20.0
dash-core-components==0.18.0
dash-html-components==0.8.0
dash-renderer==0.11.2
decorator==4.2.1
Delorean==1.0.0
Django==2.0.4
django-cors-headers==2.2.0
django-filter==1.1.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
eth-abi==1.0.0
eth-account==0.1.0a2
eth-hash==0.1.2
eth-keyfile==0.5.1
eth-keys==0.2.0b3
eth-rlp==0.1.0
eth-testrpc==1.3.3
eth-utils==1.0.3
ethereum==1.6.1
facepy==1.0.9
Flask==0.12.2
Flask-Babel==0.11.2
Flask-Cache==0.13.1
Flask-Caching==1.3.3
Flask-Compress==1.4.0
Flask-DebugToolbar==0.10.1
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
Flask-Table==0.5.0
future==0.16.0
hexbytes==0.1.0
html5lib==0.9999999
humanize==0.5.1
idna==2.6
ipython-genutils==0.2.0
itsdangerous==0.24
itypes==1.1.0
Jinja2==2.10
json-rpc==1.10.8
json2html==1.2.1
jsonschema==2.6.0
jupyter-core==4.4.0
lru-dict==1.1.6
Markdown==2.6.11
MarkupSafe==1.0
marshmallow==2.15.0
matplotlib==2.1.2
multidict==4.0.0
mysqlclient==1.3.12
nbformat==4.4.0
numpy==1.14.0
pandas==0.22.0
pbkdf2==1.3
plaid-python==2.2.1
plotly==2.2.3
poloniex==0.4.7
protobuf==3.5.1
psycopg2==2.7.4
py-solc==2.1.0
pycares==2.3.0
pycparser==2.18
pycryptodome==3.6.1
pyethash==0.1.27
PyJWT==1.6.1
pymongo==3.6.0
pyparsing==2.2.0
Pypubsub==4.0.0
pysha3==1.0.2
python-bittrex==0.1.2
python-dateutil==2.6.1
pyti==0.2.1
pytz==2018.4
PyYAML==3.12
repoze.lru==0.7
requests==2.18.4
rlp==0.6.0
scikit-learn==0.19.1
scrypt==0.8.6
secp256k1==0.13.2
semantic-version==2.6.0
six==1.11.0
sklearn==0.0
sleekxmpp==1.3.3
speculator==1.0.2
SQLAlchemy==1.2.1
stripe==1.79.1
TA-Lib==0.4.16
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
toolz==0.9.0
traitlets==4.3.2
tzlocal==1.5.1
uritemplate==3.0.0
urllib3==1.22
web3==4.1.0
webargs==2.0.0
websockets==4.0.1
Werkzeug==0.14.1
yarl==1.1.0
